requirements.txt
asgiref==3.11.0
Django==5.2.9
django-filter==25.2
django-unfold==0.72.0
djangorestframework==3.16.1
Markdown==3.10
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.4
